fix: Fix the sidebar on the components page
