+++ web_src/src/pages/workflowv2/index.tsx
@@ -30,7 +30,6 @@ import {
   workflowKeys,
 } from "@/hooks/useWorkflowData";
 import {
-  BuildingBlock,
   CanvasEdge,
   CanvasNode,
   CanvasPage,
@@ -38,7 +37,6 @@ import {
   NodeEditData,
   SidebarData,
 } from "@/ui/CanvasPage";
-import { BuildingBlockCategory } from "@/ui/BuildingBlocksSidebar";
 import { CompositeProps, LastRunState } from "@/ui/composite";
 import { getBackgroundColorClass, getColorClass } from "@/utils/colors";
 import { filterVisibleConfiguration } from "@/utils/components";
@@ -46,7 +44,7 @@ import { formatTimeAgo } from "@/utils/date";
 import { withOrganizationHeader } from "@/utils/withOrganizationHeader";
 import { getTriggerRenderer } from "./renderers";
 import { TriggerRenderer } from "./renderers/types";
-import { mockBuildingBlockCategories } from "@/ui/CanvasPage/storybooks/buildingBlocks";
+import { buildBuildingBlockCategories } from "@/ui/buildingBlocks";
 
 export function WorkflowPageV2() {
   const { organizationId, workflowId } = useParams<{
@@ -166,96 +164,11 @@ export function WorkflowPageV2() {
     isLoading: nodeDataLoading,
   } = useCompositeNodeData(workflowId!, persistedNodesWithExecutions);
 
-  // Prepare building blocks for the sidebar
-  const buildingBlocks = useMemo(() => {
-    const liveCategories: BuildingBlockCategory[] = [
-      {
-        name: "Triggers",
-        blocks: triggers.map(
-          (t): BuildingBlock => ({
-            name: t.name!,
-            label: t.label,
-            description: t.description,
-            type: "trigger",
-            configuration: t.configuration,
-            icon: t.icon,
-            color: t.color,
-            isLive: true,
-          }),
-        ),
-      },
-      {
-        name: "Primitives",
-        blocks: components.map(
-          (c): BuildingBlock => ({
-            name: c.name!,
-            label: c.label,
-            description: c.description,
-            type: "component",
-            outputChannels: c.outputChannels,
-            configuration: c.configuration,
-            icon: c.icon,
-            color: c.color,
-            isLive: true,
-          }),
-        ),
-      },
-      {
-        name: "Components",
-        blocks: blueprints.map(
-          (b): BuildingBlock => ({
-            id: b.id,
-            name: b.name!,
-            description: b.description,
-            type: "blueprint",
-            outputChannels: b.outputChannels,
-            configuration: b.configuration,
-            icon: b.icon,
-            color: b.color,
-             isLive: true,
-          }),
-        ),
-      },
-    ];
-
-    // Merge mock building blocks with live ones while avoiding duplicates
-    // Dedupe key: `${type}:${name}`
-    const byCategory = new Map<string, { blocks: Map<string, BuildingBlock>; order: string[] }>();
-
-    const addCategoryIfMissing = (name: string) => {
-      if (!byCategory.has(name)) {
-        byCategory.set(name, { blocks: new Map(), order: [] });
-      }
-    };
-
-    const addBlocks = (categoryName: string, blocks: BuildingBlock[]) => {
-      addCategoryIfMissing(categoryName);
-      const entry = byCategory.get(categoryName)!;
-      blocks.forEach((blk) => {
-        const key = `${blk.type}:${blk.name}`;
-        if (!entry.blocks.has(key)) {
-          entry.blocks.set(key, blk);
-          entry.order.push(key);
-        }
-      });
-    };
-
-    // Seed with live categories first to prioritize real components
-    liveCategories.forEach((cat) => addBlocks(cat.name, cat.blocks));
-    // Merge in mocks
-    mockBuildingBlockCategories.forEach((cat) => addBlocks(cat.name, cat.blocks));
-
-    // Materialize back to array with stable order (live-first, then mock additions)
-    const merged: BuildingBlockCategory[] = [];
-    byCategory.forEach((value, key) => {
-      merged.push({
-        name: key,
-        blocks: value.order.map((k) => value.blocks.get(k)!).filter(Boolean),
-      });
-    });
-
-    return merged;
-  }, [triggers, components, blueprints]);
+  // Prepare building blocks for the sidebar (shared logic)
+  const buildingBlocks = useMemo(
+    () => buildBuildingBlockCategories(triggers, components, blueprints),
+    [triggers, components, blueprints],
+  );
 
   const { nodes, edges } = useMemo(() => {
     // Don't prepare data until everything is loaded
